pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = [
    "setuptools",
    "packaging>=21.3",
    "twine",
    "vermin>=1.6.0",
]

[project]
name = "shark-clipper"
version = "0.0.1"
description = "A tool for getting clips of shark fins from full videos."
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}

classifiers = [
    'License :: OSI Approved :: MIT License',
    'Programming Language :: Python :: 3.8',
]

dependencies = [
    "Pillow>=10.1",
]

[project.urls]
Homepage = "https://github.com/linqs/shark-clipper"
Repository = "https://github.com/linqs/shark-clipper"
